+++ shoebox/app/com/keepit/module/CommonModule.scala
@@ -1,0 +1,142 @@
+package com.keepit.module
+
+import akka.actor.ActorSystem
+import com.google.inject.Provides
+import com.google.inject.Singleton
+import com.google.inject.multibindings.Multibinder
+import com.keepit.common.actor.ActorPlugin
+import com.keepit.common.analytics._
+import com.keepit.common.cache.MemcachedCacheModule
+import com.keepit.common.controller.FortyTwoServices
+import com.keepit.common.db.slick.Database
+import com.keepit.common.healthcheck.{HealthcheckPluginImpl, HealthcheckPlugin}
+import com.keepit.common.logging.Logging
+import com.keepit.common.mail.{MailSenderPluginImpl, MailSenderPlugin, PostOffice}
+import com.keepit.common.net.HttpClient
+import com.keepit.common.net.HttpClientImpl
+import com.keepit.common.social.{SocialGraphPluginImpl, SocialGraphPlugin}
+import com.keepit.inject.{FortyTwoModule, AppScoped}
+import com.keepit.model.SliderHistoryTracker
+import com.keepit.scraper.ScraperConfig
+import com.keepit.scraper.{HttpFetcherImpl, HttpFetcher}
+import com.keepit.search._
+import com.mongodb.casbah.MongoConnection
+import com.tzavellas.sse.guice.ScalaModule
+import java.io.File
+import java.net.InetAddress
+import play.api.Play.current
+
+class CommonModule extends ScalaModule with Logging {
+
+  def configure() {
+    install(new FortyTwoModule)
+    install(new MemcachedCacheModule)
+    install(new S3Module)
+
+    bind[ActorSystem].toProvider[ActorPlugin].in[AppScoped]
+    bind[MailSenderPlugin].to[MailSenderPluginImpl].in[AppScoped]
+    bind[SocialGraphPlugin].to[SocialGraphPluginImpl].in[AppScoped]
+
+    bind[PersistEventPlugin].to[PersistEventPluginImpl].in[AppScoped]
+
+    val listenerBinder = Multibinder.newSetBinder(binder(), classOf[EventListenerPlugin])
+    listenerBinder.addBinding().to(classOf[KifiResultClickedListener])
+    listenerBinder.addBinding().to(classOf[UsefulPageListener])
+    listenerBinder.addBinding().to(classOf[SliderShownListener])
+  }
+
+  @Singleton
+  @Provides
+  def searchConfigManager(expRepo: SearchConfigExperimentRepo, db: Database): SearchConfigManager = {
+    val optFile = current.configuration.getString("index.config").map(new File(_).getCanonicalFile).filter(_.exists)
+    new SearchConfigManager(optFile, expRepo, db)
+  }
+
+  @Singleton
+  @Provides
+  def resultClickTracker: ResultClickTracker = {
+    val conf = current.configuration.getConfig("result-click-tracker").get
+    val numHashFuncs = conf.getInt("numHashFuncs").get
+    val syncEvery = conf.getInt("syncEvery").get
+    val dirPath = conf.getString("dir").get
+    val dir = new File(dirPath).getCanonicalFile()
+    if (!dir.exists()) {
+      if (!dir.mkdirs()) {
+        throw new Exception(s"could not create dir $dir")
+      }
+    }
+    ResultClickTracker(dir, numHashFuncs, syncEvery)
+  }
+
+  @Singleton
+  @Provides
+  def clickHistoryTracker: ClickHistoryTracker = {
+    val conf = current.configuration.getConfig("click-history-tracker").get
+    val filterSize = conf.getInt("filterSize").get
+    val numHashFuncs = conf.getInt("numHashFuncs").get
+    val minHits = conf.getInt("minHits").get
+
+    ClickHistoryTracker(filterSize, numHashFuncs, minHits)
+  }
+
+  @Singleton
+  @Provides
+  def browsingHistoryTracker: BrowsingHistoryTracker = {
+    val conf = current.configuration.getConfig("browsing-history-tracker").get
+    val filterSize = conf.getInt("filterSize").get
+    val numHashFuncs = conf.getInt("numHashFuncs").get
+    val minHits = conf.getInt("minHits").get
+
+    BrowsingHistoryTracker(filterSize, numHashFuncs, minHits)
+  }
+
+  @Provides
+  @AppScoped
+  def actorPluginProvider: ActorPlugin = new ActorPlugin("shoebox-actor-system")
+
+  @Provides
+  def httpClientProvider: HttpClient = new HttpClientImpl()
+
+  @Provides
+  @AppScoped
+  def healthcheckProvider(system: ActorSystem, postOffice: PostOffice, services: FortyTwoServices): HealthcheckPlugin = {
+    val host = InetAddress.getLocalHost().getCanonicalHostName()
+    new HealthcheckPluginImpl(system, host, postOffice, services)
+  }
+
+  @Singleton
+  @Provides
+  def scraperConfig: ScraperConfig = ScraperConfig()
+
+  @Singleton
+  @Provides
+  def httpFetcher: HttpFetcher = {
+    new HttpFetcherImpl(
+      userAgent = "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_8_2) AppleWebKit/537.17 (KHTML, like Gecko) Chrome/24.0.1309.0 Safari/537.17",
+      connectionTimeout = 30000,
+      soTimeOut = 30000
+    )
+  }
+
+  @Singleton
+  @Provides
+  def mongoEventStore(): MongoEventStore = {
+    current.configuration.getString("mongo.events.server").map { server =>
+      val mongoConn = MongoConnection(server)
+      val mongoDB = mongoConn(current.configuration.getString("mongo.events.database").getOrElse("events"))
+      new MongoEventStoreImpl(mongoDB)
+    }.get
+  }
+
+  @Singleton
+  @Provides
+  def sliderHistoryTracker: SliderHistoryTracker = {
+    val conf = current.configuration.getConfig("slider-history-tracker").get
+    val filterSize = conf.getInt("filterSize").get
+    val numHashFuncs = conf.getInt("numHashFuncs").get
+    val minHits = conf.getInt("minHits").get
+
+    SliderHistoryTracker(filterSize, numHashFuncs, minHits)
+  }
+
+}
